requirements.txt
# CLI Framework
typer[all]==0.9.0

# Terminal UI and Formatting
rich==13.7.0
questionary==2.0.1
textual==0.47.1
alive-progress==3.1.5

# System Monitoring
psutil==5.9.6

# Networking
requests==2.31.0

# Data Processing
pandas==2.1.4

# Configuration
pyyaml==6.0.1

# Additional utilities for enhanced functionality
python-dateutil==2.8.2
